refactor data for bubbles and table
